Extract generation of `ChainUpdate`s from the ChainSync client test
Mostly, this involves copying code from the ChainSync client test to
new, dedicated modules.

This includes extracting the `Schedule` utilities, where `Schedule` is
now a newtype instead of a type synonym.
