+++ ouroboros-consensus-test/test-infra/Main.hs
@@ -3,6 +3,8 @@ module Main (main) where
 import           Test.Tasty
 
 import qualified Test.ThreadNet.Util.Tests (tests)
+import qualified Test.Util.ChainUpdates.Tests (tests)
+import qualified Test.Util.Schedule.Tests (tests)
 import qualified Test.Util.Split.Tests (tests)
 
 main :: IO ()
@@ -12,5 +14,7 @@ tests :: TestTree
 tests =
   testGroup "test-infra"
   [ Test.ThreadNet.Util.Tests.tests
+  , Test.Util.ChainUpdates.Tests.tests
+  , Test.Util.Schedule.Tests.tests
   , Test.Util.Split.Tests.tests
   ]
